requirements.txt
aiohttp>=3.8.4
aiosignal>=1.3.1
async-timeout>=4.0.2
attrs>=22.2.0
certifi>=2022.12.7
cffi>=1.15.1
charset-normalizer>=3.1.0
discord.py>=2.2.2
docker>=6.0.1
frozenlist>=1.3.3
idna>=3.4
multidict>=6.0.4
packaging>=23.0
pycparser>=2.21
PyNaCl>=1.5.0
python-dotenv>=1.0.0
requests>=2.28.2
toml>=0.10.2
urllib3>=1.26.15
websocket-client>=1.5.1
yarl>=1.8.2
